make listing Kodi compatible
I'm not really happy with this. It seems really like Kodi should
update it's parsing to be less strict. But, of course it's the
real world.

I just worry this is a gateway to hell where people will want
exceptions for other scrappers.

Note that the lister accepts media types by setting that `Accept:`
header

    curl -H "Accept: application/json" http://localhost:8080
    curl -H "Accept: text/plain" http://localhost:8080